% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_Covs}
\alias{add_Covs}
\title{Add Covariates and Group indices for predators}
\usage{
add_Covs(Groups = "", Covariates = "")
}
\arguments{
\item{Groups}{Index of group membership for each predator, one (named) column per grouping variable. Can be a file path to a csv file with the data or a data frame.}

\item{Covariates}{Covariate values for each predator, one (named) column per covariate. Can be a file path to a csv file with the data or a data frame.}
}
\description{
Files must be in .csv format
}
\details{
Use \code{\link{simulation}} to simulate and write these files to inspect the file structure.
}
\examples{
Cov_path <- system.file("extdata", "Simdata_Covariates.csv", package="fastinR")
Group_path<- system.file("extdata", "Simdata_Groups.csv", package="fastinR")
add_Covs(Groups=Group_path,Covariates=Cov_path)
}
\author{
Philipp Neubauer
}
\references{
Neubauer,.P. and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{add_FA}},\code{\link{add_SI}},\code{\link{run_MCMC}},\code{\link{simulation}}
}

