% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIO.R
\name{add_FA}
\alias{add_FA}
\title{Add Fatty Acid profile data for predators and prey items
    
Files must be in .csv format.}
\usage{
add_FA(
  FA.predators = NULL,
  FA.preys = NULL,
  fat.conts = "",
  Conv.Coeffs.mean = "",
  Conv.Coeffs.var = "",
  FC.mean = 1,
  FC.var = 1,
  CC.mean = 1,
  CC.var = 1,
  datas = NULL,
  LN.par = F
)
}
\arguments{
\item{FA.predators}{A data frame or a path to a csv file with the predator index/names (first column) and fatty acid profiles (1 row pey predator), with fatty acids named across the first row}

\item{FA.preys}{A data frame or a path to a csv file with the prey names/sample id (first column) and fatty acid profiles (1 row pey prey item), with fatty acids names across the first row}

\item{fat.conts}{A data frame or a path to a csv file with the prey fat contents, as (columnwise) mean and variance per prey species or specified for each prey sample for the main analysis, in that case the first column is the prey sample id id and the second column is the individual sample's fat content}

\item{Conv.Coeffs.mean}{A data frame or a path to a csv file with the prey specific conversion coefficient means: Prey names (first column) and an n x P matrix for n preys and P fatty acids}

\item{Conv.Coeffs.var}{A data frame or a path to a csv file with the prey specific conversion coefficient variances, dimensions as for the means}

\item{FC.mean}{optional - if no prey or sample specific fat content means are supplied in a fat.conts file, prey specific coefficients can be entered here using R's c(FC_1,FC_2,...) notation.}

\item{FC.var}{optional - if no prey or sample specific fat content variances are supplied in a fat.conts file, prey specific coefficients can be entered here using R's c(FC_1,FC_2,...) notation.}

\item{CC.mean}{optional - if no prey specific fractionation coefficiants are supplied via Conv.Coeffs.mean, CC.mean can provide either a global (single) mean coefficient or fatty acid specific mean coefficients using R's c(FA_1,FA_2,...) notation for ALL fatty acids.}

\item{CC.var}{optional - if no prey specific fractionation coefficiants are supplied via Conv.Coeffs.mean, CC.var can provide either a global (single) coefficient variance or fatty acid specific coefficient variances using R's c(FA_1,FA_2,...) notation for ALL fatty acids.}

\item{datas}{a data structure as produced by \code{\link{add_SI}}, needed if fatty acids and stable isotopes are added sequentially.}

\item{LN.par}{- are fat content means and variances given as log-normal parameters or sample mean and variance?}
}
\description{
Add Fatty Acid profile data for predators and prey items
    
Files must be in .csv format.
}
\details{
Use \code{\link{simulation}} to simulate and write these files to inspect the file structure.
}
\examples{
FA.predators <- system.file("extdata", "Simdata_FA_preds.csv", package="fastinR")
FA.preys <- system.file("extdata", "Simdata_FA_preys.csv", package="fastinR")
Conv.Coeffs.mean <- system.file("extdata", "Simdata_FA_cc_means.csv", package="fastinR")
Conv.Coeffs.var <- system.file("extdata", "Simdata_FA_cc_var.csv", package="fastinR")
fat.conts <- system.file("extdata", "Simdata_fat_cont.csv", package="fastinR")
dats <- add_FA(FA.predators=FA.predators,FA.preys=FA.preys,fat.conts=fat.conts,Conv.Coeffs.mean=Conv.Coeffs.mean,Conv.Coeffs.var=Conv.Coeffs.var)
}
\references{
Neubauer,.P. and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{add_SI}},\code{\link{add_Covs}},\code{\link{select_vars}},\code{\link{run_MCMC}},\code{\link{simulation}}
}
\author{
Philipp Neubauer
}
