% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{cdist}
\alias{cdist}
\title{Aitchison distance  of elements in a row wise composition matrix}
\usage{
cdist(mat)
}
\arguments{
\item{mat}{A dataframe of row-wise compositions}
}
\value{
A distance matrix
}
\description{
Aitchison distance  of elements in a row wise composition matrix
}
