% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{clo}
\alias{clo}
\title{Closure function for compositions}
\usage{
clo(x)
}
\arguments{
\item{x}{a compositional vector or dataframe of row-wise compositions}
}
\value{
A vector who's elements sum to one, or dataframe with rows summing to one
}
\description{
Implements the sum-to-unit closure for vectors and dataframes (by row).
}
