% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{clr}
\alias{clr}
\title{Centered log ratio (clr) transformation for compositions}
\usage{
clr(x)
}
\arguments{
\item{x}{a compositional vector or dataframe of row-wise compositions}
}
\value{
an clr transformed vector or dataframe
}
\description{
Implements the clr transformation for vectors and dataframes
}
