% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{dataplot}
\alias{dataplot}
\title{Plot Non-metric multidimensional scaling (NMDS) or (ilr transformed) diet data}
\usage{
dataplot(datas = NULL, trymax = 100, keep = T, new_plot = T)
}
\arguments{
\item{datas}{Diet data produced by \code{\link{add_FA}} and/or \code{\link{add_SI}}}

\item{trymax}{The maximum number of NMDS starting points.}

\item{keep}{Logical - Should the MDS results be kept for future plotting? Avoids re-running the metaMDS.}

\item{new_plot}{Logical - Should the NMDS be re-run? Defautls to true - if a previous NMDS plot was saved with the keep option, this option allows this plot to be re-drawn.}
}
\description{
Diet data produced by \code{\link{add_FA}} and/or \code{\link{add_SI}} is projected onto 2 dimesions using NMDS, where Fatty Acid data is first ilr transformed. When SI and FA data are present, then Euclidean distances of ilr(FA) and SI data are added to produce NMDS and the plot.
}
\references{
Neubauer,.P. and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{add_FA}},\code{\link{add_SI}},\code{\link{select_vars}},\code{\link{var_select_plot}},\code{\link{run_MCMC}}
}
\author{
Philipp Neubauer
}
