% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{datas}
\alias{datas}
\title{Simulated Fatty Acid and Stable Isotope example}
\description{
Simulated dataset for examples in help files, the simulation included covariates and grouped predator diets for illustration.
}
\keyword{datasets}

