% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCMC_setup.R
\name{diags}
\alias{diags}
\title{Raftery-Lewis and Gelman-Rubin diagnostics for MCMC chains from \code{\link{run_MCMC}}. Multiple chains are needed for the latter type of diagnostic.}
\usage{
diags(MCMCout = NULL, accuracy = 0.01, proba = 0.95, quant = 0.025)
}
\arguments{
\item{MCMCout}{An object produced by \code{\link{run_MCMC}}}

\item{accuracy}{Accuracy with which parameters are to be estiamted}

\item{proba}{probability with which estiamtes are within the itnerval [quant,1-quant]}

\item{quant}{interval within which to estiamte the parameter}
}
\description{
Raftery-Lewis and Gelman-Rubin diagnostics for MCMC chains from \code{\link{run_MCMC}}. Multiple chains are needed for the latter type of diagnostic.
}
\details{
The function allows to diagnose if a given MCMC run was likely long enough to have produced reliable estiamtes. The number of iterations and a thinning interval are suggested - if these are significantly larger than the parameters used to produce MCMCout, then the chains should be re-run with the suggested parameters.
}
\author{
Philipp Neubauer
}
\references{
Neubauer,.P. and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{run_MCMC}}
}

