% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{gmean}
\alias{gmean}
\title{Geometric mean of a vector or columns of a dataframe}
\usage{
gmean(x)
}
\arguments{
\item{x}{A vector or dataframe of row-wise compositions}
}
\value{
The geometric mean of a vector or columnwise geometric means for a dataframe
}
\description{
Geometric mean of a vector or columns of a dataframe
}

