% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensitivity.R
\docType{data}
\name{good_prop}
\alias{good_prop}
\title{Simulated Fatty Acid and Stable Isotope example}
\description{
Original simulated proportions for sensitivity simulations
}
\keyword{datasets}

