% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{ilr_base}
\alias{ilr_base}
\title{Ilr basis calucaltion}
\usage{
ilr_base(D)
}
\arguments{
\item{D}{Dimension of the composition}
}
\value{
The base for isometric transform
}
\description{
Ilr basis calucaltion
}

