% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{kappas}
\alias{kappas}
\title{Condition number of the source matrix as a function of a variable index}
\usage{
kappas(mat, vars)
}
\arguments{
\item{mat}{a compositional vector or dataframe of row-wise compositions}

\item{vars}{A vector of indices specifying the order in which to add columns (fatty acids)}
}
\value{
A \code{l=length(vars)} vector with condition numbers of \code{mat[vars[1:i]]} for \code{i=1:l}
}
\description{
Condition number of the source matrix as a function of a variable index
}

