% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grooming.R
\name{var_select_plot}
\alias{var_select_plot}
\title{Plot cosine distance and matrix condition number for variable selection}
\usage{
var_select_plot(prey.mat, prey.ix)
}
\arguments{
\item{prey.mat}{(row-wise) dataframe of prey compositions}

\item{prey.ix}{Indexes rows of \code{prey.mat}}
}
\value{
Two plots with 1) relative contributions of Fatty Acids to source separation and 2) matrix condition number.
}
\description{
Cosine distance based separation on CAP axes and matrix condition number are graphed as a function of Fatty Acids, where fatty acids are ordered by their relative contribution to Canonical axes in a CAP (Canonical Analysis of Principal Coordinates).
}
\examples{
 \dontrun{
# load simulated example
data('Sim')
preys <- datas$datas.FA$preys
prey.ix <- datas$prey.ix
var_select_plot(preys,prey.ix)
}
}
\references{
Neubauer.P. and Jensen, O.P. (in prep)
#' @seealso \code{\link{add_FA}},\code{\link{select_vars}},\code{\link{run_MCMC}},\code{\link{simulation}}
}
\author{
Philipp Neubauer
}
