% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Com_funcs.R
\name{adist}
\alias{adist}
\title{Cosine distance of elements in a row wise composition matrix}
\usage{
adist(mat)
}
\arguments{
\item{mat}{A dataframe of row-wise compositions}
}
\value{
A distance matrix
}
\description{
Cosine distance of elements in a row wise composition matrix
}
