% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{multiplot}
\alias{multiplot}
\title{Plot density or denstrip plots from multiple MCMC runs}
\usage{
multiplot(MCMCouts, save = "fastinR_Multiplot", types = c("violin", "strip",
  "density"))
}
\arguments{
\item{MCMCouts}{A (named) list of objects produced by  \code{\link{run_MCMC}}}

\item{save}{Either a string to be used as prefix for saved plots, or FALSE for disabling saving to file.}

\item{types}{A string to select the type of plot, either 'strip' for a density strip (needs denstrip package), and/or 'violin' for a violin plot and 'density' for dnesity plots (both use ggplot2).}
}
\description{
A density/denstrip plot produced for multiple MCMC runs from \code{\link{run_MCMC}}. Useful to compare across runs with different data subsets or data types. The analyses needs to be of the same type, for now only Analysis.Type = Population.Proportions in \code{\link{run_MCMC}} works.
}
\details{
If plots are saved they are not drawn at the same time.
}
\references{
Neubauer.P. and Jensen, O.P. (in prep)
}
\seealso{
\code{\link{run_MCMC}},\code{\link{diags}}
}
\author{
Philipp Neubauer
}
